+++ documentation/release-notes.markdown
@@ -7,7 +7,9 @@
   * Always apply the MIME type of a file when uploading to S3.
   * Speed up previewing a site with dynamic generation by only re-reading
     source files that have changed, instead of everything each time.
+  * Add a [generator for compiling SASS][sass] into CSS
 
+[sass]: https://flourish.readthedocs.io/en/latest/api-flourish-generators-sass/
 
 ## 0.7 — 16 June 2020